+++ R/KMeansTrainer.R
@@ -1,12 +1,14 @@
 #' K-Means Trainer
-#' @description Trains am unsupervised K-Means algorithm. It borrows mini-batch k-means function from
-#' ClusterR r package which is written in c++, hence it is quite fast.
+#' @description Trains a unsupervised K-Means clustering algorithm. It borrows mini-batch k-means function from
+#' ClusterR package written in c++, hence it is quite fast.
 #' @format \code{\link{R6Class}} object.
 #' @section Usage:
 #' For usage details see \bold{Methods, Arguments and Examples} sections.
 #' \preformatted{
-#' kmt = KMeansTrainer$new(n_estimators, max_features, max_depth, min_node_size, criterion,classification, class_weights, verbose, seed,always_split)
-#' bst$fit(X_train, y_train)
+#' kmt = KMeansTrainer$new(clusters, batch_size = 10, num_init=1, max_iters=100, init_fraction=1,
+#'                         initializer = "kmeans++", early_stop_iter = 10, verbose=FALSE, centroids=NULL,
+#'                         tol = 1e-04, tol_optimal_init=0.3, seed=1, max_clusters=NA)
+#' bst$fit(X_train, y_train=NULL)
 #' prediction <- bst$predict(X_test)
 #' }
 #' @section Methods:
@@ -18,7 +20,7 @@
 #' @section Arguments:
 #' \describe{
 #'  \item{params}{for explanation on parameters, please refer to the documentation of MiniBatchKMeans function in clusterR package \url{https://CRAN.R-project.org/package=ClusterR}}
-#'  \item{find_optimal}{Used to find the optimal number of cluster during \code{fit} method. To use this, make sure the value for /code{max_cluster} > 0.}
+#'  \item{find_optimal}{Used to find the optimal number of cluster during \code{fit} method. To use this, make sure the value for max_clusters > 0.}
 #' }
 #' @export
 #' @examples
@@ -38,7 +40,7 @@ KMeansTrainer <- R6Class("KMeansTrainer", public = list(
     initializer = "kmeans++",
     early_stop_iter=10,
     verbose=FALSE,
-    CENTROIDS=NULL,
+    centroids=NULL,
     tol = 1e-04,
     tol_optimal_init = 0.3,
     seed = 1,
@@ -53,24 +55,25 @@ KMeansTrainer <- R6Class("KMeansTrainer", public = list(
                           initializer = "kmeans++",
                           early_stop_iter = 10,
                           verbose=FALSE,
-                          CENTROIDS=NULL,
+                          centroids=NULL,
                           tol = 1e-04,
                           tol_optimal_init=0.3,
                           seed=1,
                           max_clusters=NA){
-        self$clusters <- clusters
-        self$batch_size <- batch_size
-        self$num_init <- 1
-        self$max_iters <- max_iters
-        self$init_fraction <- init_fraction
-        self$initializer <- initializer
-        self$early_stop_iter <- early_stop_iter
-        self$verbose <- verbose
-        self$CENTROIDS <- CENTROIDS
-        self$tol <- tol
-        self$tol_optimal_init <- tol_optimal_init
-        self$seed <- seed
-        self$max_clusters <- max_clusters
+        if(!(missing(clusters))) self$clusters <- clusters
+        if(!(missing(batch_size))) self$batch_size <- batch_size
+        if(!(missing(num_init))) self$num_init <- 1
+        if(!(missing(max_iters))) self$max_iters <- max_iters
+        if(!(missing(init_fraction))) self$init_fraction <- init_fraction
+        if(!(missing(initializer))) self$initializer <- initializer
+        if(!(missing(early_stop_iter))) self$early_stop_iter <- early_stop_iter
+        if(!(missing(verbose))) self$verbose <- verbose
+        if(!(missing(centroids))) self$centroids <- centroids
+        if(!(missing(tol))) self$tol <- tol
+        if(!(missing(tol_optimal_init)))
+            self$tol_optimal_init <- tol_optimal_init
+        if(!(missing(seed))) self$seed <- seed
+        if(!(missing(max_clusters))) self$max_clusters <- max_clusters
 
     },
 
@@ -81,7 +84,8 @@ KMeansTrainer <- R6Class("KMeansTrainer", public = list(
             stop("X should be a matrix or a data frame")
 
         if(isTRUE(find_optimal)){
-            message('Finding optimal number of clusters based on variance explained')
+            message('Finding optimal number of clusters
+                    based on variance explained')
             f <- Optimal_Clusters_KMeans(X, max_clusters = self$max_clusters)
             self$clusters <- which.max(f[-1])+1
         }
@@ -96,7 +100,7 @@ KMeansTrainer <- R6Class("KMeansTrainer", public = list(
                                       ,initializer = self$initializer
                                       ,early_stop_iter = self$early_stop_iter
                                       ,verbose = self$verbose
-                                      ,CENTROIDS = self$CENTROIDS
+                                      ,CENTROIDS = self$centroids
                                       ,tol=self$tol
                                       ,tol_optimal_init = self$tol_optimal_init
                                       ,seed = self$seed)
@@ -107,4 +111,5 @@ KMeansTrainer <- R6Class("KMeansTrainer", public = list(
     })
 
 )
+