+++ R/bm25.R
@@ -2,7 +2,7 @@
 #'
 #' @description BM25 stands for Best Matching 25. It is widely using for ranking documents and a preferred method than TF*IDF scores.
 #' It is used to find the similar documents from a corpus, given a new document. It is popularly used in information retrieval systems.
-#' This implementation ses multiple cores for faster and parallel computation.
+#' This implementation uses multiple cores for faster and parallel computation.
 #' @format \code{\link{R6Class}} object.
 #' @section Usage:
 #' For usage details see \bold{Methods, Arguments and Examples} sections.
@@ -117,4 +117,7 @@ bm25 <- R6::R6Class("bm25", public = list(
         return(names(aa[order(unlist(aa), decreasing = T)][seq(topn)]))
     }
 ))
+
+
+