+++ NEWS.md
@@ -1,3 +1,9 @@
+# superml 0.2.0
+
+* Added new trainers: SVM, NaiveBayes
+* Renamed gridsearch, randomsearch trainers to CV
+* Fixed documentation
+
 # superml 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.